[TS-1866] Added UtilsTest class
